[REFACOTR] Move all ToC related code to a seperate folder
Now it will be easy to add a depreaction and to extract the code out of
the repo.
